Move AlterationTypeConstants out of ResultsViewPageStore
