+++ src/pages/resultsView/ResultsViewPageStore.ts
@@ -241,6 +241,7 @@ import { makeUniqueColorGetter } from '../../shared/components/plots/PlotUtils';
 import ComplexKeyMap from '../../shared/lib/complexKeyDataStructures/ComplexKeyMap';
 import { getSuffixOfMolecularProfile } from 'shared/lib/molecularProfileUtils';
 import {
+    AlterationTypeConstants,
     CLINICAL_ATTRIBUTE_FIELD_ENUM,
     CLINICAL_ATTRIBUTE_ID_ENUM,
     GENETIC_PROFILE_FIELD_ENUM,
@@ -305,19 +306,6 @@ type Optional<T> =
 
 const DEFAULT_RPPA_THRESHOLD = 2;
 const DEFAULT_Z_SCORE_THRESHOLD = 2;
-
-export const AlterationTypeConstants = {
-    MUTATION_EXTENDED: 'MUTATION_EXTENDED',
-    COPY_NUMBER_ALTERATION: 'COPY_NUMBER_ALTERATION',
-    MRNA_EXPRESSION: 'MRNA_EXPRESSION',
-    PROTEIN_LEVEL: 'PROTEIN_LEVEL',
-    FUSION: 'FUSION',
-    GENESET_SCORE: 'GENESET_SCORE',
-    METHYLATION: 'METHYLATION',
-    GENERIC_ASSAY: 'GENERIC_ASSAY',
-    STRUCTURAL_VARIANT: 'STRUCTURAL_VARIANT',
-    MUTATION_UNCALLED: 'MUTATION_UNCALLED',
-};
 
 export const AlterationTypeDisplayConstants = {
     COPY_NUMBER_ALTERATION: 'CNA',
