+++ src/pages/resultsView/plots/PlotsTab.spec.tsx
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
 import { AxisMenuSelection } from 'pages/resultsView/plots/PlotsTab';
-import { AlterationTypeConstants } from 'pages/resultsView/ResultsViewPageStore';
+import { AlterationTypeConstants } from 'shared/constants';
 
 describe('PlotsTab', () => {
     describe('controls', () => {
