+++ src/pages/resultsView/plots/QuickPlots.ts
@@ -4,7 +4,7 @@ import {
     NONE_SELECTED_OPTION_STRING_VALUE,
     NONE_SELECTED_OPTION_LABEL,
 } from './PlotsTab';
-import { AlterationTypeConstants } from '../ResultsViewPageStore';
+import { AlterationTypeConstants } from 'shared/constants';
 import { SpecialAttribute } from 'shared/cache/ClinicalDataCache';
 import { CLIN_ATTR_DATA_TYPE } from './PlotsTabUtils';
 import { GenericAssayTypeConstants } from 'shared/lib/GenericAssayUtils/GenericAssayConfig';
