+++ src/pages/staticPages/tools/oncoprinter/OncoprinterImportUtils.spec.ts
@@ -5,14 +5,12 @@ import {
     getOncoprinterGeneticInput,
     getOncoprinterHeatmapInput,
 } from './OncoprinterImportUtils';
-import {
-    AlterationTypeConstants,
-    DataTypeConstants,
-} from '../../../resultsView/ResultsViewPageStore';
+import { DataTypeConstants } from '../../../resultsView/ResultsViewPageStore';
 import { SpecialAttribute } from '../../../../shared/cache/ClinicalDataCache';
 import { ONCOPRINTER_VAL_NA } from './OncoprinterClinicalAndHeatmapUtils';
 import { IHeatmapTrackSpec } from '../../../../shared/components/oncoprint/Oncoprint';
 import { PUTATIVE_DRIVER } from 'shared/lib/StoreUtils';
+import { AlterationTypeConstants } from 'shared/constants';
 
 describe('OncoprinterImportUtils', () => {
     describe('generateUniqueLabel', () => {
