+++ src/shared/lib/oql/oqlfilter.spec.ts
@@ -18,11 +18,11 @@ import _ from 'lodash';
 import { assert } from 'chai';
 import sinon from 'sinon';
 import {
-    AlterationTypeConstants,
     CustomDriverNumericGeneMolecularData,
     AnnotatedMutation,
     AnnotatedStructuralVariant,
 } from '../../../pages/resultsView/ResultsViewPageStore';
+import { AlterationTypeConstants } from 'shared/constants';
 
 // This file uses type assertions to force functions that use overly specific
 // Swagger-generated types as parameters to accept mocked literals believed to
